피드백 관련 내용 수정사항
