+++ CodeStarterCamp_Week1/main.swift
@@ -8,25 +8,3 @@
 
 import Foundation
 
-
-func creamBody(cream: Int) {
-    for _ in 1 ... cream {
-        print("***********")
-    }
-}
-
-func namooStick(stick: Int) {
-    
-    for _ in 1 ... stick {
-        print("    | |")
-    }
-    
-}
-
-
-func MakeIcecream(cream:Int , stick: Int) {
-    creamBody(cream: cream)
-    namooStick(stick: stick)
-}
-
-MakeIcecream(cream: 10, stick: 4)
